mentalwell-flask/requirements.txt
flask==3.0.3
flask-jwt-extended==4.6.0
google-generativeai==0.7.2
flask-cors==4.0.1
python-dotenv==1.0.1
gunicorn==22.0.0
